Move proc macro code out of the "new" folder
