+++ proc-macros/src/render_client.rs
@@ -1,4 +1,6 @@
-use super::{RpcDescription, RpcMethod, RpcSubscription};
+use crate::RpcDescription;
+use crate::rpc_macro::{RpcMethod, RpcSubscription};
+
 use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
